30121-muresan-laurentiu-isp-labs-30121_muresan_laurentiu_isp_labs/isp-lab-6-2023-master/src/main/java/isp/lab6/exercise2/VehicleRegistrySystem.java
package isp.lab6.exercise2;


import java.util.ArrayList;
import java.util.List;

public class VehicleRegistrySystem {
    private List<Vehicle> vehicles = new ArrayList();

    public List<Vehicle> getVehicles() {
        return vehicles;
    }

    public void addVehicle(Vehicle vehicle) {
        vehicles.add(vehicle);
    }

    public void removeVehicle(String VIN) {
        vehicles.remove(VIN);
    }

    public boolean checkVehicle(Vehicle vehicle) {
        for (Vehicle v : vehicles) {
            if (v.getVIN() == vehicle.getVIN()) {
                return true;
            }
        }
        return false;
    }

    public void displayVehicles(List<Vehicle> vehicles) {
        System.out.println(vehicles);
    }
}

